Display the results in a sheet instead of a window.
Currently the results are displayed in a window, which is nice, except for the following problem: the window stops working if you close it using the OK button, and will eventually also stop working if you close it using the red “X” button.

The OK button problem is fixable, but as far as I can tell, there is no reliable way to fix the red “X” problem at this time. See https://github.com/fileability/chocolat-public/issues/1171.

Using a sheet fixes these problems. It is not as good as a separate window would be, but is the next best thing.